fix(cli): address review — drop dead-end org picker, fix Hub auth key, add 401 hint
- Remove org-app picker listing: AppRelease has no pullable image ref and the
  in-repo cloud proto lacks GetPullCredentials, so selections were a dead-end.
  Internal-registry installs still work via explicit image ref + credentials.
- dockerConfigAuth: consult the canonical https://index.docker.io/v1/ key for
  docker.io (the real Docker CLI never stores creds under 'docker.io').
- Surface a credentials hint when a pull fails with a 401/403-style error.
- Reject an empty derived app id with a friendly --name suggestion.
